Increase version and refactor code
The version of "strava-client-rs" was increased to 0.2.1. The code in README.md, Cargo.toml, and lib.rs was refactored - specifically the unnecessary code "use strava_client_rs::api::athlete;" was removed. On top of that, a new struct and supportive functions were added in the activities.rs module to handle parameters for activities more efficiently.